refactor: change a bunch of proptype filenames
