+++ packages/article-image/src/article-image-prop-types.js
