+++ packages/article-image/src/article-image.js
@@ -6,7 +6,7 @@ import ArticleImage from "./article-image.base";
 import {
   articleImagePropTypes,
   articleImageDefaultPropTypes
-} from "./article-image-proptypes";
+} from "./article-image-prop-types";
 
 const styles = StyleSheet.create({
   primaryContainer: {
