+++ packages/brightcove-video/src/brightcove-player.android.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 
 import BrightcoveVideo from "./brightcove-player-helper";
 
-import propTypes from "./brightcove-player.proptypes";
+import propTypes from "./brightcove-player-prop-types";
 import defaults from "./brightcove-player.defaults";
 
 function withNativeCommand(WrappedComponent) {
