+++ packages/card/src/card-prop-types.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import { ViewPropTypes } from "react-native";
-import { sharedPropTypes, sharedDefaultProps } from "./card-shared-proptypes";
+import { sharedPropTypes, sharedDefaultProps } from "./card-shared-prop-types";
 
 const { style: ViewPropTypesStyle } = ViewPropTypes;
